change percision of high price items

+++ Server/ItemPrices.cs
@@ -554,9 +554,9 @@ namespace Coflnet.Sky.Core
                     new
                     {
                         End = item.Key,
-                        Avg = (int)item.Average(a => ((int)a.HighestBidAmount) / a.Count),
-                        Max = (int)item.Max(a => ((int)a.HighestBidAmount) / a.Count),
-                        Min = (int)item.Min(a => ((int)a.HighestBidAmount) / a.Count),
+                        Avg = item.Average(a => (a.HighestBidAmount) / a.Count),
+                        Max = item.Max(a => (a.HighestBidAmount) / a.Count),
+                        Min = item.Min(a => (a.HighestBidAmount) / a.Count),
                         Count = item.Sum(a => a.Count)
                     }).ToListAsync();
